Fixed issue 1
There was not enough delay between animation start and stop in case of
rapid light change, leading to assertion fail in xTaskTemrinate because
of an uninitialized task structure.

Solution: Check if the task has been fully initialized before trying to
terminate it, and add a delay of 100ms after checking if the animation
is running in MonitorTask.